move logos to own folder
